File Parser / Resolver
The parser can now take a single thrift file and follow all its
includes and parse them as well. The output is stored in a FileGroup,
which is returned.

Added a resolver to handle resolution of struct references. As
references are added to the FileGroup, they are placed in the resolver
and after parsing is complete, added to the resolutions field in the
FileGroup.