Added full thrift testing framework
Thrift tests can now create erlang files using the thrift compiler.
When this happens, records are generated in special modules as are
conversion functions to facilitate testing.

This uncovered bugs with the deserializer, which I commented out until
they're fixes. It also revealed bugs with serializing container types,
which will be fixed in a subsequent PR.