Group compiler options under a :thrift key
This change moves all compiler options under a single top-level `thrift`
configuration key that contains a Keyword list of individual options.
`thrift_files` and `thrift_output` are now just `files` and `output`
within this sub-list.

The goal of this change is to tidy up the top-level configuration space
and provide for further expansion of our compiler option set.