Include our package version in the manifest header
This gives us additional control over when to consider existing
artifacts as "stale" and in need of regeneration. At the moment, we
require a strict version string match, but as we mature in accordance
with semver guidelines, we can switch to a more fine-grained approach
based on Version.compare/2.

We don't need to bump the manifest version in this case because ...

1. Reads against existing manifests will safely fail and result in
   desirable rebuild behavior.
2. The manifest stuff is brand new, and it's unlike many folks are
   using it yet.