Implement Exception.message/1 for exceptions
We generate Elixir exception structs via defexception for Thrift
exception types. That macro always provides an exception/1
implementation, but message/1 is only generated if the exception is
defined with a :message field. When it doesn't, the generated code
produces a warning because the full Exception behaviour hasn't been
implemented.

This change generates a message/1 implementation for all of our Thrift
exception structs. It returns the Kernel.inspect/2 string representation
of the exception, which is both simple and unambiguous.

I considered only overriding message/1 when the exception struct doesn't
naturally include its own :message field. I prefer the simplicity and
consistency of the current approach because it doesn't presume anything
about the struct's layout, but this is an easy thing to revisit should
users expect different behavior.