Upgrade to dialyxir 0.5.0 (#247)
This is a significant dialyxir release because it now uses the :dialyzer
module directly instead of spawning a separate VM process in which to
run the dialyzer command line process. More details here:

https://github.com/jeremyjh/dialyxir/wiki/Upgrading-to-0.5

Fortunately for us, we get all of the benefits of this release without
having to make any specific changes to our project configuration.