Fixed Global Resolution... again
Fixes #253

The problem with the last fix is that generation is recursive. This
means we should be cleaning up each nested context's local lookups when
we move to a deeper context. Otherwise, we're relying on map iteration
order to override a particular value. If the value comes after the one
wer're inserting, we're out of luck.

This PR starts from a known good state with no overrides, and on each
subsequent generation, builds a new resolution map.