Delegate response handling to client implementation (#282)
This reduces the amount of generated code by doing the bulk of the
response processing in the common client implementation module.

It also highlights an oddity of the way we handle `void` service calls.
Even though those calls are successful, they run through the "error"
path because we don't currently have a way to differentiate between
`nil` (void) and `nil` (missing) "success" field values in the unpacked
response structure. "Reply" messages are only supposed to have either a
"success" field or an exception representation, so we'll need to revisit
our reply implementation before we can improve the success path here.