Handle reserved keywords (#295)
* Handle reserved keywords

[Similarly to apache/thrift](https://github.com/apache/thrift/blob/19baeefd8c38d62085891d7956349601f79448b3/compiler/cpp/src/thrift/thriftl.ll#L273), we disallow keywords being used in the Thrift IDL
However the keywords restriction only applies to exact-casing, so for example we must support "AND" as a valid enum. When down-casing this value, it conflicts with the Elixir keyword "and", and so using some macro magic we can get around this limitation.
Additionally, since we down-case method names too, similar logic applies when generating a client. The server implementation did not require any changes.

Fixes: #294

* Remove stray IO.inspect

* Add an 'error parsing' test case and fix typo