Support multiple destination module name casings
Previously, we expected the "struct name" portion of the input name to
be a suitable module name string (for Module.concat/1). Lowercased names
failed this assumption because they're not valid aliases.

This change reworks the destination module name processing to ensure
that the "struct name" string starts with an initial capitalize letter.

Fixes #297