Remove the client-side retry logic
In addition to the moderate state-management complexity introduced by
this feature, we've also found that implementing message-level retry
logic directly within a Thrift client to not be ideal. Instead, retries
are better implemented by either application logic (which has complete
understanding of the message payloads, idempotency, etc.) or closer to
the wire level, which can make stronger guarantees about whether the
original message was or was not delivered.