Support comment-based namespace declarations
Thrift uses the `namespace` keyword to define a language's namespace.

    namespace elixir Thrift.Namespace

Unfortunately, the Apache Thrift compiler will produce a warning on this
line because it doesn't recognize `elixir` as a supported language.
While that warning is benign, it can be annoying. For that reason, you
can also specify your Elixir namespace as a "magic" namespace comment:

    #@namespace elixir Thrift.Namespace

This alternate syntax is borrowed from Scrooge, which uses the same
trick for defining `scala` namespaces.