Remove public Thrift.Parser.parse/2 function
This is a utility function that is only used by this module's tests, so
relocate it there.

It combines two operations: parsing the document and extracting a
subcomponent based on a list of keys (the "path").

If a parsing error occurs, the function dies instead of returning the
error, which makes this a risky function to use in a real application.
Changing the return type would fix that, but I think a better solution
is to remove this function and perhaps add just the get_in/2-style
functionality to the Thrift.AST module if a need arises.