Remove the atomify/1 conversion function (#434)
This function existed to safely deal with `nil` values, but there's only
one case where this is needed: when a service doesn't "extend" another.

Instead, we now use `List.to_atom/1` directly and convert the 'extends'
value to an atom in the parser. This might be further revised if we
decide that names should be represented as a list of atoms instead of
dotted atom names (per #425).