Store FileGroup namespace values as strings
We previously stored a mapping of module_name -> %Namespace{}, but we
only need the namespace's value to resolve the destination module.

Just storing the namespace values directly (in their native string form)
simplifies this code, including the "default namespace" logic (because
it no longer needs to be represented as a full %Namespace{}).