Adopt the Mix.Task.Compiler behaviour
This is a small change that moves our compile.thrift task to the (new in
Elixir 1.6) Mix.Task.Compiler behaviour.

This will ultimately let us report diagnostics (warnings, errors, etc.)
in a standardized way, but we'll need to do some follow-up work to
better capture parser and generator errors first.