parse_file_group/2 now returns a list of errors (#450)
Because this function parses a tree of files (the root file plus all of
its recursively included files), return a separate error tuple for each
error that occurred.

FileParseError still exists but only captures the first error in the
list. I intend to remove this exception soon, so this is fine for now.

Also, while I'm here, move parser_test.exs so that it mirrors the
parser.ex location.