Return diagnostics from our compile.thrift task (#460)
This improves our support for the Mix.Task.Compiler behaviour by
returning a list of diagnostics representing any parser errors that
occurred while processing the input files.

Generator errors are not yet converted to diagnostics. That will come
next.

Errors are still printed to the shell, but they are now derived from the
Mix.Task.Compiler.Diagnostic structure. This provides a nice abstraction
and presented the opportunity to change the message format to better
match the output of the Elixir compiler.

Lastly, the unit tests have been rewritten to use a new MixTest.Case
template. Mix shell output is now routed through Mix.Shell.Process so we
no longer need to use ExUnit.CaptureIO, and the common fixture-related
helper routines are now shared.