Validate SSL configuration before starting server
If SSL certs are not available, then every SSL connection will fail. We should
check for this condition at startup before opening the port. That way if a new
node joins a cluster and its SSL configuration is not ready, it won't try to
start accepting connections that could have gone to a healthy host.