Add connect_func option to server
We'd like to do application-level validation on a client's TLS cert. Currently
it lives within the connection process state, which is not accessible from a
handler function.

In order to make it accessible, this diff allows an optional connect_func
callback to be provided when starting a Thrift server. It will be called after
a TCP or SSL connection is established with the socket as its single parameter.
Given the socket, we can retrieve the client cert using :inet.peercert/1, and
either do validation immediately or stash the cert in the process dictionary
for validation later in a handler function.