Support THeader protocol
THeader is an optional wrapper for Thrift RPC messages that allows for requests
to include out-of-band headers, useful for things like distributed tracing in
service oriented architectures.

This change only implements THeader support on the server side. We can
auto-detect the protocol on a per-request basis based on the first two bytes.
THeader starts with 0x0FFF, whereas binary starts with 0x8001.