Un-nest generated BinaryProtocol modules
A change in Elixir 1.13 prevents a nested module from referencing a struct
defined by the parent module. We use this pattern in generated BinaryProtocol
modules, which reference the struct defined by their parent, which corresponds
to a Thrift struct.

Example from the tests:

  defmodule Servers.Binary.Framed.IntegrationTest.IdAndName do
    @moduledoc false
    _ = "Auto-generated Thrift struct server_test.IdAndName"
    _ = "1: i64 id"
    _ = "2: string name"
    defstruct id: nil, name: nil
    @type t :: %__MODULE__{}
    def new do
      %__MODULE__{}
    end
    defmodule BinaryProtocol do
      @moduledoc false
      def deserialize(binary) do
        deserialize(binary, %Servers.Binary.Framed.IntegrationTest.IdAndName{})
      end

This diff eliminates the circular dependency by simply moving all of the
serialize/deserialize implementation out of BinaryProtocol into the parent.

This will break client code that references the BinaryProtocol modules
directly. However, the fix is a fairly straightforward search-and-replace.